Feat: 기본 설정, Router 설정 Remove: 필요없는 파일 삭제 Rename: GlobalStyle 이동 Chore: react-router-dom 설치 Chore: tsconfig에 baseURL 절대 경로 설정, 배포시 작동안할경우 수정 필요
components 폴더 생성하여 임의로 파일 넣어두었습니다.
