+++ src/components/todo/ToDoInput.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ToDoInput = () => {
+  return <div>ToDoInput</div>;
+};
+
+export default ToDoInput;
